jsdom: deprecate TS 3.3
It's no longer supported on Definitely Typed
