+++ types/jsdom/package.json
@@ -2,11 +2,6 @@
     "private": true,
     "types": "index",
     "typesVersions": {
-        "<=3.3": {
-            "*": [
-                "ts3.3/*"
-            ]
-        },
         "<=3.4": {
             "*": [
                 "ts3.4/*"
