--- types/jsdom/ts3.3/tslint.json
@@ -1,1 +1,0 @@
-{ "extends": "dtslint/dt.json" }
